requirements.txt
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.1
Click==8.1.8
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
requests==2.32.3
stripe==14.0.1
toml==0.10.2
urllib3==2.4.0
Werkzeug==3.1.3
gunicorn==23.0.0
Flask-SQLAlchemy==3.1.1
Authlib==1.6.0
google-auth==2.40.1
Flask-Migrate==4.1.0
pytz==2025.2
psycopg2-binary==2.9.10
slack-sdk>=3.39.0
slack-bolt>=1.27.0
APScheduler>=3.10.0
paramiko>=3.0.0
PyYAML>=6.0
pyairtable>=2.0.0
beautifulsoup4>=4.12.0
astral>=3.2
anthropic>=0.40.0
